split auth code by runtime: agnostic provider data and user code out of server/auth
