+++ waspc/data/Generator/templates/sdk/wasp/auth/index.ts
@@ -6,4 +6,4 @@ export {
 } from './user.js'
 
 // PUBLIC API
-export { type AuthUser } from '../server/auth/user.js'
+export { type AuthUser } from './user.js'
