+++ waspc/data/Generator/templates/sdk/wasp/auth/types.ts
@@ -1,2 +1,1 @@
-// todo(filip): turn into a proper import/path
-export type { ProviderName } from '../server/auth/utils.js'
+export type { ProviderName } from './providerData.js'
