+++ waspc/data/Generator/templates/sdk/wasp/server/auth/utils.ts
@@ -11,45 +11,29 @@ import { Prisma } from '@prisma/client';
 
 import { throwValidationError } from '../../auth/validation.js'
 
+import {
+  type ProviderId,
+  type ProviderName,
+  type PossibleProviderData,
+  providerDataHasPasswordField,
+} from '../../auth/providerData.js'
+
 import { type UserSignupFields, type PossibleUserFields } from './providers/types.js'
 
-// PUBLIC API
-export type EmailProviderData = {
-  hashedPassword: string;
-  isEmailVerified: boolean;
-  emailVerificationSentAt: string | null;
-  passwordResetSentAt: string | null;
-}
-
-// PUBLIC API
-export type UsernameProviderData = {
-  hashedPassword: string;
-}
-
-// PUBLIC API
-export type OAuthProviderData = {}
-
-// PRIVATE API
-/**
- * This type is used for type-level programming e.g. to enumerate
- * all possible provider data types.
- * 
- * The keys of this type are the names of the providers and the values
- * are the types of the provider data.
- */
-export type PossibleProviderData = {
-  email: EmailProviderData;
-  username: UsernameProviderData;
-  discord: OAuthProviderData;
-  slack: OAuthProviderData;
-  google: OAuthProviderData;
-  keycloak: OAuthProviderData;
-  github: OAuthProviderData;
-  microsoft: OAuthProviderData;
-}
-
-// PUBLIC API
-export type ProviderName = keyof PossibleProviderData
+// Runtime-agnostic provider data code, re-exported here because it's part of
+// the server-side auth API surface (e.g. through `wasp/server/auth`).
+export {
+  createProviderId,
+  normalizeProviderUserId,
+  getProviderData,
+  getProviderDataWithPassword,
+  type ProviderId,
+  type ProviderName,
+  type PossibleProviderData,
+  type EmailProviderData,
+  type UsernameProviderData,
+  type OAuthProviderData,
+} from '../../auth/providerData.js'
 
 // PRIVATE API
 export const contextWithUserEntity = {
@@ -65,59 +49,6 @@ export const authConfig = {
 }
 
 // PUBLIC API
-/**
- * ProviderId uniquely identifies an auth identity e.g. 
- * "email" provider with user id "test@test.com" or
- * "google" provider with user id "1234567890".
- * 
- * We use this type to avoid passing the providerName and providerUserId
- * separately. Also, we can normalize the providerUserId to make sure it's
- * consistent across different DB operations.
- */
-export type ProviderId = {
-  providerName: ProviderName;
-  providerUserId: string;
-}
-
-// PUBLIC API
-export function createProviderId(providerName: ProviderName, providerUserId: string): ProviderId {
-  return {
-    providerName,
-    providerUserId: normalizeProviderUserId(providerName, providerUserId),
-  }
-}
-
-// PRIVATE API
-export function normalizeProviderUserId(providerName: ProviderName, providerUserId: string): string {
-  switch (providerName) {
-    case 'email':
-    case 'username':
-      return providerUserId.toLowerCase();
-    case 'google':
-    case 'github':
-    case 'discord':
-    case 'keycloak':
-    case 'slack':
-    case 'microsoft':
-      return providerUserId;
-    /*
-      Why the default case?
-      In case users add a new auth provider in the user-land.
-      Users can't extend this function because it is private.
-      If there is an unknown `providerName` in runtime, we'll
-      return the `providerUserId` as is.
-
-      We want to still have explicit OAuth providers listed
-      so that we get a type error if we forget to add a new provider
-      to the switch statement.
-    */
-    default:
-      providerName satisfies never;
-      return providerUserId;
-  }
-}
-
-// PUBLIC API
 export async function findAuthIdentity(providerId: ProviderId): Promise<{= authIdentityEntityUpper =} | null> {
   return prisma.{= authIdentityEntityLower =}.findUnique({
     where: {
@@ -129,7 +60,7 @@ export async function findAuthIdentity(providerId: ProviderId): Promise<{= authI
 // PUBLIC API
 /**
  * Updates the provider data for the given auth identity.
- * 
+ *
  * This function performs data sanitization and serialization.
  * Sanitization is done by hashing the password, so this function
  * expects the password received in the `providerDataUpdates`
@@ -227,7 +158,7 @@ export async function deleteUserByAuthId(authId: string): Promise<{ count: numbe
 // will make it harder for an attacker to determine
 // if a user exists or not.
 // NOTE: Attacker measuring time to response can still determine
-// if a user exists or not. We'll be able to avoid it when 
+// if a user exists or not. We'll be able to avoid it when
 // we implement e-mail sending via jobs.
 export async function doFakeWork(): Promise<unknown> {
   const timeToWork = Math.floor(Math.random() * 1000) + 1000;
@@ -307,32 +238,6 @@ export async function validateAndGetUserFields(
 }
 
 // PUBLIC API
-export function getProviderData<PN extends ProviderName>(
-  providerData: string,
-):  Omit<PossibleProviderData[PN], 'hashedPassword'> {
-  return sanitizeProviderData(getProviderDataWithPassword(providerData));
-}
-
-// PUBLIC API
-export function getProviderDataWithPassword<PN extends ProviderName>(
-  providerData: string,
-): PossibleProviderData[PN] {
-  // NOTE: We are letting JSON.parse throw an error if the providerData is not valid JSON.
-  return JSON.parse(providerData);
-}
-
-function sanitizeProviderData<PN extends ProviderName>(
-  providerData: PossibleProviderData[PN],
-): Omit<PossibleProviderData[PN], 'hashedPassword'> {
-  if (providerDataHasPasswordField(providerData)) {
-    const { hashedPassword, ...rest } = providerData;
-    return rest;
-  } else {
-    return providerData;
-  }
-}
-
-// PUBLIC API
 export async function sanitizeAndSerializeProviderData<PN extends ProviderName>(
   providerData: PossibleProviderData[PN],
 ): Promise<string> {
@@ -356,13 +261,6 @@ async function ensurePasswordIsHashed<PN extends ProviderName>(
   }
 
   return data;
-}
-
-
-function providerDataHasPasswordField(
-  providerData: PossibleProviderData[keyof PossibleProviderData],
-): providerData is { hashedPassword: string } {
-  return 'hashedPassword' in providerData;
 }
 
 // PRIVATE API
