+++ waspc/data/Generator/templates/sdk/wasp/server/utils.ts
@@ -2,7 +2,7 @@
 import { Response, RequestHandler } from 'express'
 
 {=# isAuthEnabled =}
-import { type AuthUserData } from './auth/user.js'
+import { type AuthUserData } from '../auth/user.js'
 {=/ isAuthEnabled =}
 
 // This is explicitly how Express expects extensions to their
