+++ waspc/data/Generator/templates/sdk/wasp/server/webSocket/index.ts
@@ -5,7 +5,7 @@ import { EventsMap, DefaultEventsMap } from '@socket.io/component-emitter'
 
 import { prisma } from '../index.js'
 {=# isAuthEnabled =}
-import { type AuthUser } from '../auth/user.js'
+import { type AuthUser } from '../../auth/user.js'
 {=/ isAuthEnabled =}
 
 {=& userWebSocketFn.importStatement =}
